Use JAS version 2.6.5765

+++ symja_android_library/matheclipse-core/src/main/java/org/matheclipse/core/expression/BigIntegerSym.java
@@ -128,10 +128,7 @@ public class BigIntegerSym extends AbstractIntegerSym {
 		return valueOf(newnum);
 	}
 
-	/**
-	 * @return
-	 */
-	public int bitLength() {
+	public long bitLength() {
 		return fBigIntValue.bitLength();
 	}
